requirements.txt
docx2txt==0.8
nltk==3.8.1
pandas==2.2.2
pdfminer.six==20231228
pymongo==4.7.2
python-dotenv==1.0.1
Sastrawi==1.0.1
Sastrawi==1.0.1
scikit_learn==1.5.0
